refactor: move `components` module into folder and split `EntityIId` into its own file
