+++ src/components/mod.rs
@@ -1,11 +1,12 @@
 //! [Component]s and [Bundle]s used by the plugin.
+mod entity_iid;
+pub use entity_iid::EntityIid;
 
 pub use crate::ldtk::EntityInstance;
 use crate::ldtk::{LayerInstance, Type};
 use bevy::prelude::*;
 
 use std::{
-    borrow::Cow,
     collections::HashSet,
     ops::{Add, AddAssign, Mul, MulAssign, Sub, SubAssign},
 };
@@ -19,60 +20,6 @@ use crate::{
 };
 
 use bevy_ecs_tilemap::tiles::{TileBundle, TilePos};
-
-/// [Component] added to all [LdtkEntity]s by default.
-///
-/// The `iid` stored in this component can be used to uniquely identify LDtk entities within an [LdtkAsset].
-#[derive(Clone, Debug, Default, Deref, Hash, Eq, PartialEq, Component, FromReflect, Reflect)]
-#[reflect(Component, Default, Debug)]
-pub struct EntityIid(Cow<'static, str>);
-
-impl EntityIid {
-    pub fn new(iid: impl Into<Cow<'static, str>>) -> Self {
-        let iid = iid.into();
-        EntityIid(iid)
-    }
-
-    #[inline(always)]
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl From<&str> for EntityIid {
-    #[inline(always)]
-    fn from(value: &str) -> Self {
-        EntityIid::new(value.to_owned())
-    }
-}
-
-impl From<String> for EntityIid {
-    #[inline(always)]
-    fn from(value: String) -> Self {
-        EntityIid::new(value)
-    }
-}
-
-impl AsRef<str> for EntityIid {
-    #[inline(always)]
-    fn as_ref(&self) -> &str {
-        &self.0
-    }
-}
-
-impl From<&EntityIid> for String {
-    #[inline(always)]
-    fn from(value: &EntityIid) -> String {
-        value.as_str().to_owned()
-    }
-}
-
-impl From<EntityIid> for String {
-    #[inline(always)]
-    fn from(value: EntityIid) -> String {
-        value.0.into_owned()
-    }
-}
 
 /// [Component] added to any `IntGrid` tile by default.
 ///
